// server/src/generated/prisma/models/ChatHistory.ts
export type ChatHistoryModel = runtime.Types.Result.DefaultSelection<Prisma.$ChatHistoryPayload>
// ...
export type $ChatHistoryPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "ChatHistory"
  objects: {
    app: Prisma.$AppPayload<ExtArgs>
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: bigint
    message: string
    messageType: $Enums.ChatMessageType
    appId: bigint
    userId: bigint
    createTime: Date
    updateTime: Date
    isDelete: boolean
  }, ExtArgs["result"]["chatHistory"]>
  composites: {}
}